requirements.txt
a2wsgi==1.10.7
aiobotocore==2.15.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.13.2
anyio==4.4.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-sqlite==3.8.2
apispec==6.6.1
argcomplete==3.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
avro-python3==1.10.2
babel==2.16.0
backoff==2.2.1
blinker==1.8.2
botocore==1.35.16
cachelib==0.9.0
cattrs==24.1.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.1
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
email-validator==1.3.1
executing==2.1.0
Flask==2.2.5
Flask-AppBuilder==4.3.3
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.0
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kafka-python==2.0.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
minio==7.2.8
more-itertools==10.5.0
msgspec==0.18.6
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pendulum==2.1.2
platformdirs==4.2.2
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.18
Pygments==2.18.0
PyJWT==2.9.0
pyspark==3.5.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
s3fs==2024.9.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
spotipy==2.24.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.5
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==2.3.8
wrapt==1.16.0
WTForms==3.1.2
yarl==1.11.1
